reafactor: SettingView 내 섹션 분리
